+++ common/src/main/java/com/microsoft/identity/common/internal/cache/SharedPreferencesBrokerApplicationMetadataCache.java
@@ -23,12 +23,11 @@
 package com.microsoft.identity.common.internal.cache;
 
 import android.content.Context;
-import android.content.SharedPreferences;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import android.text.TextUtils;
 
-import com.google.gson.Gson;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import com.google.gson.reflect.TypeToken;
 import com.microsoft.identity.common.internal.logging.Logger;
 
@@ -39,6 +38,7 @@ import java.util.List;
 import java.util.Set;
 
 public class SharedPreferencesBrokerApplicationMetadataCache
+        extends SharedPreferencesSimpleCacheImpl<BrokerApplicationMetadata>
         implements IBrokerApplicationMetadataCache {
 
     private static final String TAG = SharedPreferencesBrokerApplicationMetadataCache.class.getSimpleName();
@@ -46,21 +46,9 @@ public class SharedPreferencesBrokerApplicationMetadataCache
     private static final String DEFAULT_APP_METADATA_CACHE_NAME = "com.microsoft.identity.app-meta-cache";
 
     private static final String KEY_CACHE_LIST = "app-meta-cache";
-    private static final String EMPTY_ARRAY = "[]";
-
-    private final SharedPreferences mSharedPrefs;
-
-    private final Gson mGson = new Gson();
 
     public SharedPreferencesBrokerApplicationMetadataCache(@NonNull final Context context) {
-        Logger.verbose(
-                TAG + "::ctor",
-                "Init"
-        );
-        mSharedPrefs = context.getSharedPreferences(
-                DEFAULT_APP_METADATA_CACHE_NAME,
-                Context.MODE_PRIVATE
-        );
+        super(context, DEFAULT_APP_METADATA_CACHE_NAME, KEY_CACHE_LIST);
     }
 
     @Override
@@ -182,142 +170,8 @@ public class SharedPreferencesBrokerApplicationMetadataCache
     }
 
     @Override
-    public synchronized boolean insert(@NonNull final BrokerApplicationMetadata metadata) {
-        final String methodName = ":insert";
-
-        final Set<BrokerApplicationMetadata> allMetadata = new HashSet<>(getAll());
-        Logger.verbose(
-                TAG + methodName,
-                "Existing metadata contained ["
-                        + allMetadata.size()
-                        + "] elements."
-        );
-
-        allMetadata.add(metadata);
-
-        Logger.verbose(
-                TAG + methodName,
-                "New metadata set size: ["
-                        + allMetadata.size()
-                        + "]"
-        );
-
-        final String json = mGson.toJson(allMetadata);
-
-        Logger.verbose(
-                TAG + methodName,
-                "Writing cache entry."
-        );
-
-        final boolean success = mSharedPrefs.edit().putString(KEY_CACHE_LIST, json).commit();
-
-        if (success) {
-            Logger.verbose(
-                    TAG + methodName,
-                    "Cache successfully updated."
-            );
-        } else {
-            Logger.warn(
-                    TAG + methodName,
-                    "Error writing to cache."
-            );
-        }
-
-        return success;
-    }
-
-    @Override
-    public synchronized boolean remove(@NonNull final BrokerApplicationMetadata metadata) {
-        final String methodName = ":remove";
-
-        final Set<BrokerApplicationMetadata> allMetadata = new HashSet<>(getAll());
-
-        Logger.verbose(
-                TAG + methodName,
-                "Existing metadata contained ["
-                        + allMetadata.size()
-                        + "] elements."
-        );
-
-        final boolean removed = allMetadata.remove(metadata);
-
-        Logger.verbose(
-                TAG + methodName,
-                "New metadata set size: ["
-                        + allMetadata.size()
-                        + "]"
-        );
-
-        if (!removed) {
-            // Nothing to do, wasn't cached in the first place!
-            Logger.warn(
-                    TAG + methodName,
-                    "Nothing to delete -- cache entry is missing!"
-            );
-
-            return true;
-        } else {
-            final String json = mGson.toJson(allMetadata);
-
-            Logger.verbose(
-                    TAG + methodName,
-                    "Writing new cache values..."
-            );
-
-            final boolean written = mSharedPrefs.edit().putString(KEY_CACHE_LIST, json).commit();
-
-            Logger.verbose(
-                    TAG + methodName,
-                    "Updated cache contents written? ["
-                            + written
-                            + "]"
-            );
-
-            return written;
-        }
-    }
-
-    @Override
-    public synchronized List<BrokerApplicationMetadata> getAll() {
-        final String methodName = ":getAll";
-        final String jsonList = mSharedPrefs.getString(KEY_CACHE_LIST, EMPTY_ARRAY);
-
-        final Type listType = new TypeToken<List<BrokerApplicationMetadata>>() {
+    protected Type getListTypeToken() {
+        return new TypeToken<List<BrokerApplicationMetadata>>() {
         }.getType();
-
-        final List<BrokerApplicationMetadata> result = mGson.fromJson(
-                jsonList,
-                listType
-        );
-
-        Logger.verbose(
-                TAG + methodName,
-                "Found ["
-                        + result.size()
-                        + "] cache entries."
-        );
-
-        return result;
-    }
-
-    @Override
-    public synchronized boolean clear() {
-        final String methodName = ":clear";
-
-        final boolean cleared = mSharedPrefs.edit().clear().commit();
-
-        if (!cleared) {
-            Logger.warn(
-                    TAG + methodName,
-                    "Failed to clear cache."
-            );
-        } else {
-            Logger.verbose(
-                    TAG + methodName,
-                    "Cache successfully cleared."
-            );
-        }
-
-        return cleared;
     }
 }
